third_party/WebKit/LayoutTests/inspector/network/network-filter-parser.html

<html>
<head>
<script src="../../http/tests/inspector/inspector-test.js"></script>
<script src="../../http/tests/inspector/network-test.js"></script>
<script>
function test() {
    function checkQuery(keys, query)
    {
        var suggestionBuilder = new WebInspector.FilterSuggestionBuilder(keys);
        var result = suggestionBuilder.parseQuery(query);
        InspectorTest.addResult("");
        InspectorTest.addResult("Keys: " + JSON.stringify(keys));
        InspectorTest.addResult("Query: " + query);
        InspectorTest.addResult("Result: " + JSON.stringify(result));
    }

    checkQuery([], "plain text");
    checkQuery(["single"], "single:value");
    checkQuery(["single"], "-single:value");
    checkQuery(["one", "two"], "one:foo two:bar");
    checkQuery(["one", "two"], "-one:foo two:bar");
    checkQuery(["one", "two"], "-one:foo -two:bar");
    checkQuery(["one", "two"], "one:foo -two:bar");
    checkQuery(["key"], "bar key foo");
    checkQuery(["key"], "bar key:foo");
    checkQuery(["key"], "bar key:foo baz");
    checkQuery(["key"], "bar key:foo yek:roo baz");
    checkQuery(["key"], "bar key:foo -yek:roo baz");
    checkQuery(["key"], "bar baz key:foo goo zoo");
    checkQuery(["key"], "bar key:key:foo");
    checkQuery(["key"], "bar :key:foo baz");
    checkQuery(["key"], "bar -:key:foo baz");
    checkQuery(["key"], "bar key:-foo baz");

    InspectorTest.completeTest();
}

</script>
</head>
<body onload="runTest()">
<p>Tests filter query parsing.</p>
</body>
</html>
